requirements.txt
python-telegram-bot==20.7
google-cloud-speech==2.23.0
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0